server/python/requirements.txt
langchain==0.1.0
chromadb==0.4.22
sentence-transformers==2.2.2
groq==0.4.2
python-dotenv==1.0.0
flask==3.0.2
flask-cors==4.0.0 